feat(services): added Read/List for movie entity.
